Allow the user to explicitly reset the thread pool
It's difficult to use multi-threading with fork(). The Verilator
thread pool belongs to the parent process only after fork() is called.
This commit adds a member function to allow the user explicitly
resetting the thread pool from the child process after fork().
Only after the thread pool is created, the simulation could continue.

This is dirty but I don't know other ways to do this.

One clean way is to use pthread_atfork. However, it does not accept
C++ lambda functions. We need lambdas instead of function pointers
to catch the `this` context. Therefore, this is not possible.